requirements.txt
astroid==2.4.2
Boruta==0.3
boto3==1.14.15
botocore==1.17.15
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
docutils==0.15.2
easy-sklearn==0.1.5
httpie==2.1.0
idna==2.9
imbalanced-learn==0.6.2
imblearn==0.0
inline==0.0.1
isort==4.3.21
jmespath==0.10.0
joblib==0.15.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.2.1
mccabe==0.6.1
mixed-naive-bayes==0.0.1
mlens==0.2.3
mlxtend==0.17.2
numexpr==2.7.1
numpy==1.18.4
pandas==1.0.3
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
rope==0.17.0
s3transfer==0.3.3
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
tables==3.6.1
threadpoolctl==2.0.0
toml==0.10.1
urllib3==1.25.9
wrapt==1.12.1
